Rename JsonParser to AirbrakeApi::V3::NoticeParser
Also I've added additional spec for requests with api_key set to nil.
