+++ lib/airbrake_api/params_error.rb
@@ -1,0 +1,3 @@
+module AirbrakeApi
+  class ParamsError < StandardError; end
+end
