+++ spec/lib/airbrake_api/v3/notice_parser_spec.rb
@@ -1,18 +1,43 @@
-describe JsonParser do
+describe AirbrakeApi::V3::NoticeParser do
   let(:app) { Fabricate(:app) }
 
   it 'raises error when errors attribute is missing' do
     expect {
-      JsonParser.new({}).report
-    }.to raise_error(JsonParser::ParamsError)
+      AirbrakeApi::V3::NoticeParser.new({}).report
+    }.to raise_error(AirbrakeApi::V3::NoticeParser::ParamsError)
 
     expect {
-      JsonParser.new({'errors' => []}).report
-    }.to raise_error(JsonParser::ParamsError)
+      AirbrakeApi::V3::NoticeParser.new({'errors' => []}).report
+    }.to raise_error(AirbrakeApi::V3::NoticeParser::ParamsError)
   end
 
   it 'parses JSON payload and returns ErrorReport' do
-    params = JSON.parse(<<-EOL)
+    params = build_params(api_key: app.api_key)
+
+    report = AirbrakeApi::V3::NoticeParser.new(params).report
+    notice = report.generate_notice!
+
+    expect(report.error_class).to eq('Error')
+    expect(report.message).to eq('Error: TestError')
+    expect(report.backtrace.lines.size).to eq(9)
+    expect(notice.user_attributes).to include({'Id' => 1, 'Name' => 'John Doe', 'Email' => 'john.doe@example.org', 'Username' => 'john'})
+    expect(notice.session).to include('isAdmin' => true)
+    expect(notice.params).to include('returnTo' => 'dashboard')
+    expect(notice.env_vars).to include(
+      'navigator_vendor' => 'Google Inc.',
+      'HTTP_USER_AGENT' => 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_5) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/39.0.2171.99 Safari/537.36'
+    )
+  end
+
+  it 'parses JSON payload when api_key is missing but project_id is present' do
+    params = build_params(api_key: nil, project_id: app.api_key)
+
+    report = AirbrakeApi::V3::NoticeParser.new(params).report
+    expect(report).to be_valid
+  end
+
+  def build_params(options = {})
+    JSON.parse(<<-EOL)
       {
         "notifier":{"name":"airbrake-js-v8","version":"0.3.10","url":"https://github.com/airbrake/airbrake-js"},
         "errors":[
@@ -48,22 +73,9 @@ describe JsonParser do
         "params":{"returnTo":"dashboard"},
         "environment":{"navigator_vendor":"Google Inc."},
         "session":{"isAdmin":true},
-        "key":"#{app.api_key}"
+        "key":"#{options[:api_key]}",
+        "project_id":"#{options[:project_id]}"
       }
     EOL
-
-    report = JsonParser.new(params).report
-    notice = report.generate_notice!
-
-    expect(report.error_class).to eq('Error')
-    expect(report.message).to eq('Error: TestError')
-    expect(report.backtrace.lines.size).to eq(9)
-    expect(notice.user_attributes).to include({'Id' => 1, 'Name' => 'John Doe', 'Email' => 'john.doe@example.org', 'Username' => 'john'})
-    expect(notice.session).to include('isAdmin' => true)
-    expect(notice.params).to include('returnTo' => 'dashboard')
-    expect(notice.env_vars).to include(
-      'navigator_vendor' => 'Google Inc.',
-      'HTTP_USER_AGENT' => 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_5) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/39.0.2171.99 Safari/537.36'
-    )
   end
 end
